Refactor some code into the pex_registry_fixture
